Update some adapters, viewModels, fragments

+++ app/common/src/commonJvmMain/kotlin/com/denchic45/kts/data/repository/SubmissionRepository.kt
@@ -60,7 +60,7 @@ class SubmissionRepository(
         courseId: UUID,
         workId: UUID,
         submissionId: UUID,
-        grade: Short,
+        grade: Int,
     ) = fetchResource {
         submissionsApi.gradeSubmission(courseId, workId, submissionId, grade)
     }
