+++ common/api/src/commonMain/kotlin/com/denchic45/stuiversity/api/course/work/submission/model/Grade.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 @Serializable
 data class GradeRequest(
-    val value: Short,
+    val value: Int,
 )
